Improved corner case concerning ByReference types
SharpRemote tries to avoid creating proxies for proxies which can happen when a by reference type is passed A->B->A (where A and B are endpoints). Previously, the last invocation would receive a proxy to a proxy (that would function correctly, but punish the caller by causing 2 RPC invocations).
Instead SharpRemote ensures that the 2nd invocation receives the original subject instead.