Fixed generated code jumps #57
Generated code no longer attempts a direct jump from an exception handling block to outside of it. Instead we first jump to the end of the try block (at which point a leave instruction will be posted). Apparently this is verified in AppVeyor's .NET runtime, but not in the one on my computer